+++ app/src/main/java/org/girino/frac/operators/ShipBarOperator.java
@@ -13,6 +13,11 @@ public class ShipBarOperator extends FractalOperator {
 	protected void beforeIteration(int step, Complex Z, Complex C, int maxiter) {
 		C.conjugate();
 	}
+
+	@Override
+	protected void resumeIteration(int step, Complex Z, Complex C, int maxiter) {
+		C.conjugate();
+	}
 	protected void step(int step, Complex Z, Complex C, int maxiter) {
 		Z.abs();
 		Z.square();